requirements.txt
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
numpy==1.26.2
streamlit==1.39.0
plotly==5.24.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
requests==2.31.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
openai==1.54.0
